Centralized region caching into RegionContext, so repositories can access the same cache, and avoid duplicating memoization.

+++ src/main/java/com/kntrel/mc/regionLib/persistence/sqlite/DTO.java
@@ -25,7 +25,7 @@ final class DTO {
             @Column("name") String name,
             @Column("world") String world,
             @Column("enabled") boolean enabled,
-            @Column("hierarchy") int hierarchy,
+            @Column("hierarchy") long hierarchy,
             @Column("min_x") double minX,
             @Column("min_y") double minY,
             @Column("min_z") double minZ,
